[#29] Verify stdout capturing in fork mode
With Gradle 2.0+.
